lab1-airbnb/agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
SQLAlchemy==2.0.35
pymysql==1.1.1
httpx==0.27.2

# LangChain + tools
langchain==0.2.14
langchain-community==0.2.12
tavily-python==0.3.6